move NoC functions to a separate file
